feat: split nrfutil and nrfutil device exports to avoid cyclic dependencies
